Move project, only support NET 6.
