+++ dotnet/src/dotnetcore/DynService/Cosmos/CosmosDBDatastoreHelper.cs
