+++ dotnet/src/dotnetcore/DynService/Cosmos/CosmosDBMaps.cs
