+++ dotnet/src/dotnetcore/DynService/Cosmos/CosmosDBRequestWrapper.cs
