+++ dotnet/src/dotnetcore/DynService/Cosmos/CosmosDBResponseWrapper.cs
