+++ dotnet/src/dotnetcore/DynService/Cosmos/DynService.CosmosDB.csproj
@@ -1,6 +1,6 @@
 <Project Sdk="Microsoft.NET.Sdk">
 	<PropertyGroup>
-		<TargetFrameworks>net462;net6.0</TargetFrameworks>
+		<TargetFrameworks>net6.0</TargetFrameworks>
 		<RootNamespace>GeneXus.Data.NTier</RootNamespace>
 		<AssemblyName>GeneXus.Data.DynService.CosmosDB</AssemblyName>
 		<SignAssembly>false</SignAssembly>
@@ -8,16 +8,6 @@
 		<PackageId>GeneXus.DynService.CosmosDB</PackageId>
 	</PropertyGroup>
 	
-	<PropertyGroup Condition="'$(TargetFramework)'!='net462'">
-		<DefineConstants>NETCORE</DefineConstants>
-	</PropertyGroup>
-
-	<ItemGroup Condition="'$(TargetFramework)'!='net462'">
-		<ProjectReference Include="..\..\dotnetcore\GxClasses\GxClasses.csproj" />
-	</ItemGroup>
-	<ItemGroup Condition="'$(TargetFramework)'=='net462'">
-		<ProjectReference Include="..\..\dotnetframework\GxClasses\GxClasses.csproj" />
-	</ItemGroup>
 	<ItemGroup>
 	  <Content Remove="C:\Users\sjuarez\.nuget\packages\microsoft.azure.cosmos\3.31.0\contentFiles\any\netstandard2.0\ThirdPartyNotice.txt" />
 	</ItemGroup>
@@ -27,6 +17,6 @@
 	</ItemGroup>
 
 	<ItemGroup>
-	  <ProjectReference Include="..\DynService.Core\DynService.Core.csproj" />
+	  <ProjectReference Include="..\..\..\dotnetcommon\DynService.Core\DynService.Core.csproj" />
 	</ItemGroup>
 </Project>
